requirements.txt
Django==3.2.5
django-crispy-forms==1.8.1
django-filter==2.2.0
djangorestframework==3.11.2
Markdown==3.1.1
social-auth-app-django==3.1.0
social-auth-core==3.3.3
django-extensions==2.2.5
django-taggit==1.3.0
six==1.16.0